+++ lib/src/callable/async_built_in.dart
@@ -26,6 +26,11 @@ class AsyncBuiltInCallable implements AsyncCallable {
   /// The callback to run when executing this callable.
   final Callback _callback;
 
+  /// Whether this callable could potentially accept an `@content` block.
+  ///
+  /// This can only be true for mixins.
+  final bool acceptsContent;
+
   /// Creates a function with a single [arguments] declaration and a single
   /// [callback].
   ///
@@ -52,7 +57,7 @@ class AsyncBuiltInCallable implements AsyncCallable {
   /// defined.
   AsyncBuiltInCallable.mixin(String name, String arguments,
       FutureOr<void> callback(List<Value> arguments),
-      {Object? url})
+      {Object? url, bool acceptsContent = false})
       : this.parsed(name,
             ArgumentDeclaration.parse('@mixin $name($arguments) {', url: url),
             (arguments) async {
@@ -66,7 +71,8 @@ class AsyncBuiltInCallable implements AsyncCallable {
 
   /// Creates a callable with a single [arguments] declaration and a single
   /// [callback].
-  AsyncBuiltInCallable.parsed(this.name, this._arguments, this._callback);
+  AsyncBuiltInCallable.parsed(this.name, this._arguments, this._callback,
+      {this.acceptsContent = false});
 
   /// Returns the argument declaration and Dart callback for the given
   /// positional and named arguments.