+++ lib/src/visitor/evaluate.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 4dbb0910bb0250c07a9ae7c302d9c9403d94c12a
+// Checksum: 08ee05790c0d5cce3b190278723c52dadaee1701
 //
 // ignore_for_file: unused_import
 
@@ -49,6 +49,7 @@ import '../module/built_in.dart';
 import '../parse/keyframe_selector.dart';
 import '../syntax.dart';
 import '../utils.dart';
+import '../util/character.dart';
 import '../util/map.dart';
 import '../util/multi_span.dart';
 import '../util/nullable.dart';
@@ -426,6 +427,19 @@ final class _EvaluateVisitor
         });
       }, url: "sass:meta"),
 
+      BuiltInCallable.function("module-mixins", r"$module", (arguments) {
+        var namespace = arguments[0].assertString("module");
+        var module = _environment.modules[namespace.text];
+        if (module == null) {
+          throw 'There is no module with namespace "${namespace.text}".';
+        }
+
+        return SassMap({
+          for (var (name, value) in module.mixins.pairs)
+            SassString(name): SassMixin(value)
+        });
+      }, url: "sass:meta"),
+
       BuiltInCallable.function(
           "get-function", r"$name, $css: false, $module: null", (arguments) {
         var name = arguments[0].assertString("name");
@@ -439,13 +453,31 @@ final class _EvaluateVisitor
           return SassFunction(PlainCssCallable(name.text));
         }
 
-        var callable = _addExceptionSpan(
-            _callableNode!,
-            () => _getFunction(name.text.replaceAll("_", "-"),
-                namespace: module?.text));
+        var callable = _addExceptionSpan(_callableNode!, () {
+          var normalizedName = name.text.replaceAll("_", "-");
+          var namespace = module?.text;
+          var local =
+              _environment.getFunction(normalizedName, namespace: namespace);
+          if (local != null || namespace != null) return local;
+          return _builtInFunctions[normalizedName];
+        });
         if (callable == null) throw "Function not found: $name";
 
         return SassFunction(callable);
+      }, url: "sass:meta"),
+
+      BuiltInCallable.function("get-mixin", r"$name, $module: null",
+          (arguments) {
+        var name = arguments[0].assertString("name");
+        var module = arguments[1].realNull?.assertString("module");
+
+        var callable = _addExceptionSpan(
+            _callableNode!,
+            () => _environment.getMixin(name.text.replaceAll("_", "-"),
+                namespace: module?.text));
+        if (callable == null) throw "Mixin not found: $name";
+
+        return SassMixin(callable);
       }, url: "sass:meta"),
 
       BuiltInCallable.function("call", r"$function, $args...", (arguments) {
@@ -518,7 +550,32 @@ final class _EvaluateVisitor
             configuration: configuration,
             namesInErrors: true);
         _assertConfigurationIsEmpty(configuration, nameInError: true);
-      }, url: "sass:meta")
+      }, url: "sass:meta"),
+      BuiltInCallable.mixin("apply", r"$mixin, $args...", (arguments) {
+        var mixin = arguments[0];
+        var args = arguments[1] as SassArgumentList;
+
+        var callableNode = _callableNode!;
+        var invocation = ArgumentInvocation(
+          const [],
+          const {},
+          callableNode.span,
+          rest: ValueExpression(args, callableNode.span),
+        );
+
+        var callable = mixin.assertMixin("mixin").callable;
+        var content = _environment.content;
+
+        // ignore: unnecessary_type_check
+        if (callable is Callable) {
+          _applyMixin(
+              callable, content, invocation, callableNode, callableNode);
+        } else {
+          throw SassScriptException(
+              "The mixin ${callable.name} is asynchronous.\n"
+              "This is probably caused by a bug in a Sass plugin.");
+        }
+      }, url: "sass:meta", acceptsContent: true),
     ];
 
     var metaModule = BuiltInModule("meta",
@@ -1727,40 +1784,55 @@ final class _EvaluateVisitor
     }
   }
 
-  Value? visitIncludeRule(IncludeRule node) {
-    var nodeWithSpan = AstNode.fake(() => node.spanWithoutContent);
-    var mixin = _addExceptionSpan(node,
-        () => _environment.getMixin(node.name, namespace: node.namespace));
+  /// Evaluate a given [mixin] with [arguments] and [contentCallable]
+  void _applyMixin(
+      Callable? mixin,
+      UserDefinedCallable<Environment>? contentCallable,
+      ArgumentInvocation arguments,
+      AstNode nodeWithSpan,
+      AstNode nodeWithSpanWithoutContent) {
     switch (mixin) {
       case null:
-        throw _exception("Undefined mixin.", node.span);
+        throw _exception("Undefined mixin.", nodeWithSpan.span);
 
-      case BuiltInCallable() when node.content != null:
-        throw _exception("Mixin doesn't accept a content block.", node.span);
-
+      case BuiltInCallable(acceptsContent: false) when contentCallable != null:
+        {
+          var evaluated = _evaluateArguments(arguments);
+          var (overload, _) = mixin.callbackFor(
+              evaluated.positional.length, MapKeySet(evaluated.named));
+          throw MultiSpanSassRuntimeException(
+              "Mixin doesn't accept a content block.",
+              nodeWithSpanWithoutContent.span,
+              "invocation",
+              {overload.spanWithName: "declaration"},
+              _stackTrace(nodeWithSpanWithoutContent.span));
+        }
       case BuiltInCallable():
-        _runBuiltInCallable(node.arguments, mixin, nodeWithSpan);
+        _environment.withContent(contentCallable, () {
+          _environment.asMixin(() {
+            _runBuiltInCallable(arguments, mixin, nodeWithSpanWithoutContent);
+          });
+        });
 
       case UserDefinedCallable<Environment>(
             declaration: MixinRule(hasContent: false)
           )
-          when node.content != null:
+          when contentCallable != null:
         throw MultiSpanSassRuntimeException(
             "Mixin doesn't accept a content block.",
-            node.spanWithoutContent,
+            nodeWithSpanWithoutContent.span,
             "invocation",
             {mixin.declaration.arguments.spanWithName: "declaration"},
-            _stackTrace(node.spanWithoutContent));
+            _stackTrace(nodeWithSpanWithoutContent.span));
 
       case UserDefinedCallable<Environment>():
-        var contentCallable = node.content.andThen((content) =>
-            UserDefinedCallable(content, _environment.closure(),
-                inDependency: _inDependency));
-        _runUserDefinedCallable(node.arguments, mixin, nodeWithSpan, () {
+        _runUserDefinedCallable(arguments, mixin, nodeWithSpanWithoutContent,
+            () {
           _environment.withContent(contentCallable, () {
             _environment.asMixin(() {
               for (var statement in mixin.declaration.children) {
-                _addErrorSpan(nodeWithSpan, () => statement.accept(this));
+                _addErrorSpan(
+                    nodeWithSpanWithoutContent, () => statement.accept(this));
               }
             });
           });
@@ -1769,6 +1841,20 @@ final class _EvaluateVisitor
       case _:
         throw UnsupportedError("Unknown callable type $mixin.");
     }
+  }
+
+  Value? visitIncludeRule(IncludeRule node) {
+    var mixin = _addExceptionSpan(node,
+        () => _environment.getMixin(node.name, namespace: node.namespace));
+    var contentCallable = node.content.andThen((content) => UserDefinedCallable(
+        content, _environment.closure(),
+        inDependency: _inDependency));
+
+    var nodeWithSpanWithoutContent =
+        AstNode.fake(() => node.spanWithoutContent);
+
+    _applyMixin(mixin, contentCallable, node.arguments, node,
+        nodeWithSpanWithoutContent);
 
     return null;
   }
@@ -2172,6 +2258,13 @@ final class _EvaluateVisitor
   // ## Expressions
 
   Value visitBinaryOperationExpression(BinaryOperationExpression node) {
+    if (_stylesheet.plainCss &&
+        node.operator != BinaryOperator.singleEquals &&
+        node.operator != BinaryOperator.dividedBy) {
+      throw _exception(
+          "Operators aren't allowed in plain CSS.", node.operatorSpan);
+    }
+
     return _addExceptionSpan(node, () {
       var left = node.left.accept(this);
       return switch (node.operator) {
@@ -2202,7 +2295,10 @@ final class _EvaluateVisitor
   Value _slash(Value left, Value right, BinaryOperationExpression node) {
     var result = left.dividedBy(right);
     switch ((left, right)) {
-      case (SassNumber left, SassNumber right) when node.allowsSlash:
+      case (SassNumber left, SassNumber right)
+          when node.allowsSlash &&
+              _operandAllowsSlash(node.left) &&
+              _operandAllowsSlash(node.right):
         return (result as SassNumber).withSlash(left, right);
 
       case (SassNumber(), SassNumber()):
@@ -2234,6 +2330,20 @@ final class _EvaluateVisitor
         return result;
     }
   }
+
+  /// Returns whether [node] can be used as a component of a slash-separated
+  /// number.
+  ///
+  /// Although this logic is mostly resolved at parse-time, we can't tell
+  /// whether operands will be evaluated as calculations until evaluation-time.
+  bool _operandAllowsSlash(Expression node) =>
+      node is! FunctionExpression ||
+      (node.namespace == null &&
+          const {
+            "calc", "clamp", "hypot", "sin", "cos", "tan", "asin", "acos", //
+            "atan", "sqrt", "exp", "sign", "mod", "rem", "atan2", "pow", "log"
+          }.contains(node.name.toLowerCase()) &&
+          _environment.getFunction(node.name) == null);
 
   Value visitValueExpression(ValueExpression node) => node.value;
 
@@ -2278,23 +2388,143 @@ final class _EvaluateVisitor
       SassNumber(node.value, node.unit);
 
   Value visitParenthesizedExpression(ParenthesizedExpression node) =>
-      node.expression.accept(this);
+      _stylesheet.plainCss
+          ? throw _exception(
+              "Parentheses aren't allowed in plain CSS.", node.span)
+          : node.expression.accept(this);
 
-  Value visitCalculationExpression(CalculationExpression node) {
+  SassColor visitColorExpression(ColorExpression node) => node.value;
+
+  SassList visitListExpression(ListExpression node) => SassList(
+      node.contents.map((Expression expression) => expression.accept(this)),
+      node.separator,
+      brackets: node.hasBrackets);
+
+  SassMap visitMapExpression(MapExpression node) {
+    var map = <Value, Value>{};
+    var keyNodes = <Value, AstNode>{};
+    for (var (key, value) in node.pairs) {
+      var keyValue = key.accept(this);
+      var valueValue = value.accept(this);
+
+      if (map.containsKey(keyValue)) {
+        var oldValueSpan = keyNodes[keyValue]?.span;
+        throw MultiSpanSassRuntimeException(
+            'Duplicate key.',
+            key.span,
+            'second key',
+            {if (oldValueSpan != null) oldValueSpan: 'first key'},
+            _stackTrace(key.span));
+      }
+      map[keyValue] = valueValue;
+      keyNodes[keyValue] = key;
+    }
+    return SassMap(map);
+  }
+
+  Value visitFunctionExpression(FunctionExpression node) {
+    var function = _stylesheet.plainCss
+        ? null
+        : _addExceptionSpan(
+            node,
+            () =>
+                _environment.getFunction(node.name, namespace: node.namespace));
+    if (function == null) {
+      if (node.namespace != null) {
+        throw _exception("Undefined function.", node.span);
+      }
+
+      switch (node.name.toLowerCase()) {
+        case "min" || "max" || "round" || "abs"
+            when node.arguments.named.isEmpty &&
+                node.arguments.rest == null &&
+                node.arguments.positional
+                    .every((argument) => argument.isCalculationSafe):
+          return _visitCalculation(node, inLegacySassFunction: true);
+
+        case "calc" ||
+              "clamp" ||
+              "hypot" ||
+              "sin" ||
+              "cos" ||
+              "tan" ||
+              "asin" ||
+              "acos" ||
+              "atan" ||
+              "sqrt" ||
+              "exp" ||
+              "sign" ||
+              "mod" ||
+              "rem" ||
+              "atan2" ||
+              "pow" ||
+              "log":
+          return _visitCalculation(node);
+      }
+
+      function = (_stylesheet.plainCss ? null : _builtInFunctions[node.name]) ??
+          PlainCssCallable(node.originalName);
+    }
+
+    var oldInFunction = _inFunction;
+    _inFunction = true;
+    var result = _addErrorSpan(
+        node, () => _runFunctionCallable(node.arguments, function, node));
+    _inFunction = oldInFunction;
+    return result;
+  }
+
+  Value _visitCalculation(FunctionExpression node,
+      {bool inLegacySassFunction = false}) {
+    if (node.arguments.named.isNotEmpty) {
+      throw _exception(
+          "Keyword arguments can't be used with calculations.", node.span);
+    } else if (node.arguments.rest != null) {
+      throw _exception(
+          "Rest arguments can't be used with calculations.", node.span);
+    }
+
+    _checkCalculationArguments(node);
     var arguments = [
-      for (var argument in node.arguments)
-        _visitCalculationValue(argument,
-            inMinMax: node.name == 'min' || node.name == 'max')
+      for (var argument in node.arguments.positional)
+        _visitCalculationExpression(argument,
+            inLegacySassFunction: inLegacySassFunction)
     ];
     if (_inSupportsDeclaration) {
       return SassCalculation.unsimplified(node.name, arguments);
     }
 
+    var oldCallableNode = _callableNode;
+    _callableNode = node;
+
     try {
-      return switch (node.name) {
+      return switch (node.name.toLowerCase()) {
         "calc" => SassCalculation.calc(arguments[0]),
+        "sqrt" => SassCalculation.sqrt(arguments[0]),
+        "sin" => SassCalculation.sin(arguments[0]),
+        "cos" => SassCalculation.cos(arguments[0]),
+        "tan" => SassCalculation.tan(arguments[0]),
+        "asin" => SassCalculation.asin(arguments[0]),
+        "acos" => SassCalculation.acos(arguments[0]),
+        "atan" => SassCalculation.atan(arguments[0]),
+        "abs" => SassCalculation.abs(arguments[0]),
+        "exp" => SassCalculation.exp(arguments[0]),
+        "sign" => SassCalculation.sign(arguments[0]),
         "min" => SassCalculation.min(arguments),
         "max" => SassCalculation.max(arguments),
+        "hypot" => SassCalculation.hypot(arguments),
+        "pow" =>
+          SassCalculation.pow(arguments[0], arguments.elementAtOrNull(1)),
+        "atan2" =>
+          SassCalculation.atan2(arguments[0], arguments.elementAtOrNull(1)),
+        "log" =>
+          SassCalculation.log(arguments[0], arguments.elementAtOrNull(1)),
+        "mod" =>
+          SassCalculation.mod(arguments[0], arguments.elementAtOrNull(1)),
+        "rem" =>
+          SassCalculation.rem(arguments[0], arguments.elementAtOrNull(1)),
+        "round" => SassCalculation.round(arguments[0],
+            arguments.elementAtOrNull(1), arguments.elementAtOrNull(2)),
         "clamp" => SassCalculation.clamp(arguments[0],
             arguments.elementAtOrNull(1), arguments.elementAtOrNull(2)),
         _ => throw UnsupportedError('Unknown calculation name "${node.name}".')
@@ -2303,8 +2533,53 @@ final class _EvaluateVisitor
       // The simplification logic in the [SassCalculation] static methods will
       // throw an error if the arguments aren't compatible, but we have access
       // to the original spans so we can throw a more informative error.
-      _verifyCompatibleNumbers(arguments, node.arguments);
+      if (error.message.contains("compatible")) {
+        _verifyCompatibleNumbers(arguments, node.arguments.positional);
+      }
       throwWithTrace(_exception(error.message, node.span), error, stackTrace);
+    } finally {
+      _callableNode = oldCallableNode;
+    }
+  }
+
+  /// Verifies that the calculation [node] has the correct number of arguments.
+  void _checkCalculationArguments(FunctionExpression node) {
+    void check([int? maxArgs]) {
+      if (node.arguments.positional.isEmpty) {
+        throw _exception("Missing argument.", node.span);
+      } else if (maxArgs != null &&
+          node.arguments.positional.length > maxArgs) {
+        throw _exception(
+            "Only $maxArgs ${pluralize('argument', maxArgs)} allowed, but "
+                    "${node.arguments.positional.length} " +
+                pluralize('was', node.arguments.positional.length,
+                    plural: 'were') +
+                " passed.",
+            node.span);
+      }
+    }
+
+    switch (node.name.toLowerCase()) {
+      case "calc" ||
+            "sqrt" ||
+            "sin" ||
+            "cos" ||
+            "tan" ||
+            "asin" ||
+            "acos" ||
+            "atan" ||
+            "abs" ||
+            "exp" ||
+            "sign":
+        check(1);
+      case "min" || "max" || "hypot":
+        check();
+      case "pow" || "atan2" || "log" || "mod" || "rem":
+        check(2);
+      case "round" || "clamp":
+        check(3);
+      case _:
+        throw UnsupportedError('Unknown calculation name "${node.name}".');
     }
   }
 
@@ -2345,54 +2620,47 @@ final class _EvaluateVisitor
 
   /// Evaluates [node] as a component of a calculation.
   ///
-  /// If [inMinMax] is `true`, this allows unitless numbers to be added and
+  /// If [inLegacySassFunction] is `true`, this allows unitless numbers to be added and
   /// subtracted with numbers with units, for backwards-compatibility with the
-  /// old global `min()` and `max()` functions.
-  Object _visitCalculationValue(Expression node, {required bool inMinMax}) {
+  /// old global `min()`, `max()`, `round()`, and `abs()` functions.
+  Object _visitCalculationExpression(Expression node,
+      {required bool inLegacySassFunction}) {
     switch (node) {
       case ParenthesizedExpression(expression: var inner):
-        var result = _visitCalculationValue(inner, inMinMax: inMinMax);
-        return inner is FunctionExpression &&
-                inner.name.toLowerCase() == 'var' &&
-                result is SassString &&
-                !result.hasQuotes
+        var result = _visitCalculationExpression(inner,
+            inLegacySassFunction: inLegacySassFunction);
+        return result is SassString
             ? SassString('(${result.text})', quotes: false)
             : result;
 
-      case StringExpression(text: Interpolation(asPlain: var text?)):
+      case StringExpression() when node.isCalculationSafe:
         assert(!node.hasQuotes);
-        return switch (text.toLowerCase()) {
+        return switch (node.text.asPlain?.toLowerCase()) {
           'pi' => SassNumber(math.pi),
           'e' => SassNumber(math.e),
           'infinity' => SassNumber(double.infinity),
           '-infinity' => SassNumber(double.negativeInfinity),
           'nan' => SassNumber(double.nan),
-          _ => SassString(text, quotes: false)
+          _ => SassString(_performInterpolation(node.text), quotes: false)
         };
 
-      // If there's actual interpolation, create a CalculationInterpolation.
-      // Otherwise, create an UnquotedString. The main difference is that
-      // UnquotedStrings don't get extra defensive parentheses.
-      case StringExpression():
-        assert(!node.hasQuotes);
-        return CalculationInterpolation(_performInterpolation(node.text));
-
       case BinaryOperationExpression(:var operator, :var left, :var right):
+        _checkWhitespaceAroundCalculationOperator(node);
         return _addExceptionSpan(
             node,
             () => SassCalculation.operateInternal(
-                _binaryOperatorToCalculationOperator(operator),
-                _visitCalculationValue(left, inMinMax: inMinMax),
-                _visitCalculationValue(right, inMinMax: inMinMax),
-                inMinMax: inMinMax,
+                _binaryOperatorToCalculationOperator(operator, node),
+                _visitCalculationExpression(left,
+                    inLegacySassFunction: inLegacySassFunction),
+                _visitCalculationExpression(right,
+                    inLegacySassFunction: inLegacySassFunction),
+                inLegacySassFunction: inLegacySassFunction,
                 simplify: !_inSupportsDeclaration));
 
-      case _:
-        assert(node is NumberExpression ||
-            node is CalculationExpression ||
-            node is VariableExpression ||
-            node is FunctionExpression ||
-            node is IfExpression);
+      case NumberExpression() ||
+            VariableExpression() ||
+            FunctionExpression() ||
+            IfExpression():
         return switch (node.accept(this)) {
           SassNumber result => result,
           SassCalculation result => result,
@@ -2400,68 +2668,104 @@ final class _EvaluateVisitor
           var result => throw _exception(
               "Value $result can't be used in a calculation.", node.span)
         };
+
+      case ListExpression(
+          hasBrackets: false,
+          separator: ListSeparator.space,
+          contents: [_, _, ...]
+        ):
+        var elements = [
+          for (var element in node.contents)
+            _visitCalculationExpression(element,
+                inLegacySassFunction: inLegacySassFunction)
+        ];
+
+        _checkAdjacentCalculationValues(elements, node);
+
+        for (var i = 0; i < elements.length; i++) {
+          if (elements[i] is CalculationOperation &&
+              node.contents[i] is ParenthesizedExpression) {
+            elements[i] = SassString("(${elements[i]})", quotes: false);
+          }
+        }
+
+        return SassString(elements.join(' '), quotes: false);
+
+      case _:
+        assert(!node.isCalculationSafe);
+        throw _exception(
+            "This expression can't be used in a calculation.", node.span);
+    }
+  }
+
+  /// Throws an error if [node] requires whitespace around its operator in a
+  /// calculation but doesn't have it.
+  void _checkWhitespaceAroundCalculationOperator(
+      BinaryOperationExpression node) {
+    if (node.operator != BinaryOperator.plus &&
+        node.operator != BinaryOperator.minus) {
+      return;
+    }
+
+    // We _should_ never be able to violate these conditions since we always
+    // parse binary operations from a single file, but it's better to be safe
+    // than have this crash bizarrely.
+    if (node.left.span.file != node.right.span.file) return;
+    if (node.left.span.end.offset >= node.right.span.start.offset) return;
+
+    var textBetweenOperands = node.left.span.file
+        .getText(node.left.span.end.offset, node.right.span.start.offset);
+    var first = textBetweenOperands.codeUnitAt(0);
+    var last = textBetweenOperands.codeUnitAt(textBetweenOperands.length - 1);
+    if (!(first.isWhitespace || first == $slash) ||
+        !(last.isWhitespace || last == $slash)) {
+      throw _exception(
+          '"+" and "-" must be surrounded by whitespace in calculations.',
+          node.operatorSpan);
     }
   }
 
   /// Returns the [CalculationOperator] that corresponds to [operator].
   CalculationOperator _binaryOperatorToCalculationOperator(
-          BinaryOperator operator) =>
+          BinaryOperator operator, BinaryOperationExpression node) =>
       switch (operator) {
         BinaryOperator.plus => CalculationOperator.plus,
         BinaryOperator.minus => CalculationOperator.minus,
         BinaryOperator.times => CalculationOperator.times,
         BinaryOperator.dividedBy => CalculationOperator.dividedBy,
-        _ => throw UnsupportedError("Invalid calculation operator $operator.")
+        _ => throw _exception(
+            "This operation can't be used in a calculation.", node.operatorSpan)
       };
 
-  SassColor visitColorExpression(ColorExpression node) => node.value;
+  /// Throws an error if [elements] contains two adjacent non-string values.
+  void _checkAdjacentCalculationValues(
+      List<Object> elements, ListExpression node) {
+    assert(elements.length > 1);
 
-  SassList visitListExpression(ListExpression node) => SassList(
-      node.contents.map((Expression expression) => expression.accept(this)),
-      node.separator,
-      brackets: node.hasBrackets);
+    for (var i = 1; i < elements.length; i++) {
+      var previous = elements[i - 1];
+      var current = elements[i];
+      if (previous is SassString || current is SassString) continue;
 
-  SassMap visitMapExpression(MapExpression node) {
-    var map = <Value, Value>{};
-    var keyNodes = <Value, AstNode>{};
-    for (var (key, value) in node.pairs) {
-      var keyValue = key.accept(this);
-      var valueValue = value.accept(this);
-
-      var oldValue = map[keyValue];
-      if (oldValue != null) {
-        var oldValueSpan = keyNodes[keyValue]?.span;
-        throw MultiSpanSassRuntimeException(
-            'Duplicate key.',
-            key.span,
-            'second key',
-            {if (oldValueSpan != null) oldValueSpan: 'first key'},
-            _stackTrace(key.span));
+      var previousNode = node.contents[i - 1];
+      var currentNode = node.contents[i];
+      if (currentNode
+          case UnaryOperationExpression(
+                operator: UnaryOperator.minus || UnaryOperator.plus
+              ) ||
+              NumberExpression(value: < 0)) {
+        // `calc(1 -2)` parses as a space-separated list whose second value is a
+        // unary operator or a negative number, but just saying it's an invalid
+        // expression doesn't help the user understand what's going wrong. We
+        // add special case error handling to help clarify the issue.
+        throw _exception(
+            '"+" and "-" must be surrounded by whitespace in calculations.',
+            currentNode.span.subspan(0, 1));
+      } else {
+        throw _exception('Missing math operator.',
+            previousNode.span.expand(currentNode.span));
       }
-      map[keyValue] = valueValue;
-      keyNodes[keyValue] = key;
     }
-    return SassMap(map);
-  }
-
-  Value visitFunctionExpression(FunctionExpression node) {
-    var function = _addExceptionSpan(
-        node, () => _getFunction(node.name, namespace: node.namespace));
-
-    if (function == null) {
-      if (node.namespace != null) {
-        throw _exception("Undefined function.", node.span);
-      }
-
-      function = PlainCssCallable(node.originalName);
-    }
-
-    var oldInFunction = _inFunction;
-    _inFunction = true;
-    var result = _addErrorSpan(
-        node, () => _runFunctionCallable(node.arguments, function, node));
-    _inFunction = oldInFunction;
-    return result;
   }
 
   Value visitInterpolatedFunctionExpression(
@@ -2473,14 +2777,6 @@ final class _EvaluateVisitor
         node, () => _runFunctionCallable(node.arguments, function, node));
     _inFunction = oldInFunction;
     return result;
-  }
-
-  /// Like `_environment.getFunction`, but also returns built-in
-  /// globally-available functions.
-  Callable? _getFunction(String name, {String? namespace}) {
-    var local = _environment.getFunction(name, namespace: namespace);
-    if (local != null || namespace != null) return local;
-    return _builtInFunctions[name];
   }
 
   /// Evaluates the arguments in [arguments] as applied to [callable], and
@@ -3159,7 +3455,7 @@ final class _EvaluateVisitor
   String _performInterpolation(Interpolation interpolation,
       {bool warnForColor = false}) {
     var (result, _) = _performInterpolationHelper(interpolation,
-        sourceMap: true, warnForColor: warnForColor);
+        sourceMap: false, warnForColor: warnForColor);
     return result;
   }