+++ lib/src/visitor/serialize.dart
@@ -518,13 +518,9 @@ final class _SerializeVisitor
       case Value():
         value.accept(this);
 
-      case CalculationInterpolation():
-        _buffer.write(value.value);
-
       case CalculationOperation(:var operator, :var left, :var right):
-        var parenthesizeLeft = left is CalculationInterpolation ||
-            (left is CalculationOperation &&
-                left.operator.precedence < operator.precedence);
+        var parenthesizeLeft = left is CalculationOperation &&
+            left.operator.precedence < operator.precedence;
         if (parenthesizeLeft) _buffer.writeCharCode($lparen);
         _writeCalculationValue(left);
         if (parenthesizeLeft) _buffer.writeCharCode($rparen);
@@ -534,8 +530,7 @@ final class _SerializeVisitor
         _buffer.write(operator.operator);
         if (operatorWhitespace) _buffer.writeCharCode($space);
 
-        var parenthesizeRight = right is CalculationInterpolation ||
-            (right is CalculationOperation &&
+        var parenthesizeRight = (right is CalculationOperation &&
                 _parenthesizeCalculationRhs(operator, right.operator)) ||
             (operator == CalculationOperator.dividedBy &&
                 right is SassNumber &&
@@ -565,38 +560,29 @@ final class _SerializeVisitor
       case ColorSpace.rgb || ColorSpace.hsl || ColorSpace.hwb:
         _writeLegacyColor(value);
 
-      case ColorSpace.lab || ColorSpace.oklab:
+      case ColorSpace.lab ||
+            ColorSpace.oklab ||
+            ColorSpace.lch ||
+            ColorSpace.oklch:
         _buffer
           ..write(value.space)
           ..writeCharCode($lparen);
-        _writeChannel(value.channel0OrNull);
-        if (!_isCompressed &&
-            value.space == ColorSpace.lab &&
-            !value.isChannel0Missing) {
+        if (!_isCompressed && !value.isChannel0Missing) {
+          var max = (value.space.channels[0] as LinearChannel).max;
+          _writeNumber(value.channel0 * 100 / max);
           _buffer.writeCharCode($percent);
+        } else {
+          _writeChannel(value.channel0OrNull);
         }
         _buffer.writeCharCode($space);
         _writeChannel(value.channel1OrNull);
         _buffer.writeCharCode($space);
         _writeChannel(value.channel2OrNull);
-        _maybeWriteSlashAlpha(value);
-        _buffer.writeCharCode($rparen);
-
-      case ColorSpace.lch || ColorSpace.oklch:
-        _buffer
-          ..write(value.space)
-          ..writeCharCode($lparen);
-        _writeChannel(value.channel0OrNull);
         if (!_isCompressed &&
-            value.space == ColorSpace.lch &&
-            !value.isChannel0Missing) {
-          _buffer.writeCharCode($percent);
+            !value.isChannel2Missing &&
+            value.space.channels[2].isPolarAngle) {
+          _buffer.write('deg');
         }
-        _buffer.writeCharCode($space);
-        _writeChannel(value.channel1OrNull);
-        _buffer.writeCharCode($space);
-        _writeChannel(value.channel2OrNull);
-        if (!_isCompressed && !value.isChannel2Missing) _buffer.write('deg');
         _maybeWriteSlashAlpha(value);
         _buffer.writeCharCode($rparen);
 
@@ -836,6 +822,16 @@ final class _SerializeVisitor
 
     _buffer.write("get-function(");
     _visitQuotedString(function.callable.name);
+    _buffer.writeCharCode($rparen);
+  }
+
+  void visitMixin(SassMixin mixin) {
+    if (!_inspect) {
+      throw SassScriptException("$mixin isn't a valid CSS value.");
+    }
+
+    _buffer.write("get-mixin(");
+    _visitQuotedString(mixin.callable.name);
     _buffer.writeCharCode($rparen);
   }